+++ Fireworks/ExplosionUpdater.cs
@@ -18,14 +18,14 @@ public class ExplosionUpdater(Game game)
     {
         foreach (var explosionParticle in explosion.Particles)
         {
-            explosionParticle.Age += Timer.DeltaTimeMilliseconds;
+            explosionParticle.Age += (float)Timer.DeltaTimeMilliseconds;
 
             var gravityForce = (explosionParticle.Position - game.Earth.Position).Normalize() * game.Gravity * AirResistance;
             var airResistanceDecay = -(explosionParticle.Velocity.Normalize() * AirResistance);
 
-            explosionParticle.Acceleration += (gravityForce + airResistanceDecay) / explosionParticle.Mass * Timer.DeltaTimeSeconds;
+            explosionParticle.Acceleration += (gravityForce + airResistanceDecay) / explosionParticle.Mass * (float)Timer.DeltaTimeSeconds;
             explosionParticle.Velocity += explosionParticle.Acceleration;
-            explosionParticle.Position += explosionParticle.Velocity * Timer.DeltaTimeSeconds;
+            explosionParticle.Position += explosionParticle.Velocity * (float)Timer.DeltaTimeSeconds;
         }
     }
 }
